+++ doc/release_notes.rst
@@ -10,6 +10,13 @@ Release Notes
 
 Upcoming Open-TYNDP Release
 ================
+
+**Features**
+
+* Add TYNDP hydro technologies and associated PEMMDB capacities and inflows (https://github.com/open-energy-transition/open-tyndp/pull/338).
+
+**Documentation**
+
 * Docs: Add innovation roadmap comparison, mapping existing or new developments in PyPSA-Eur and Open TYNDP to desired features for TYNDP 2026 (https://github.com/open-energy-transition/open-tyndp/pull/341)
 
 Upcoming PyPSA-Eur Release